Default to using modern syntax when Rails.version is undefined
When I run the test suite by executing `bundle exec rake` from the project root directory, I get a `undefined method `version' for Rails:Module (NoMethodError)` error on line 9 of lib/perspectives/railtie.rb. It seems like `Rails.version` is undefined, although I'm not clear on why this is the case. Rails is specified as a development dependency in the gemspec, which presumably means that it is available in the development and test environments. In any case, this commit makes changes that preserve the original behavior in Perspectives::Railtie while also defaulting to using the more modern syntax when Rails.version is undefined.